// src/app/interfaces/elemento.ts
export interface IElemento {
  id?: number;
  familia: string | null;
  tipoElemento: string | null;
  faces: string | null;
  elementos: string | null;
  maiorDimensao: string | null;
  menorDimensao: string | null;
  comprimento: string | null;
  somaLados: string | null;
  areaSuperficieUnitaria: string | null;
  areaSuperficieGeral: string | null;
  areaAdicional: string | null;
  areaTotal: string | null;
  observacoes: string | null;
}